refactor(import): change pages import rule
