+++ apps/client/src/App.tsx
@@ -1,7 +1,7 @@
 import { BROWSE_PATH, SEARCH_PATH } from "@/constants/paths";
 import { useHashLocation } from "@/hooks/use-hash-location";
-import Browse from "@/pages/browse/browse";
-import Search from "@/pages/search/search";
+import Browse from "@/pages/browse/Browse";
+import Search from "@/pages/search/Search";
 import { Route, Router } from "wouter";
 
 export function App() {
